nits: glorious nits phase 2
